requirements.txt
aiohttp==3.8.4
aiohttp-proxy==0.1.2
aiosignal==1.3.1
alicebot==0.6.1
alicebot-adapter-mirai==0.6.0.post1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
frozenlist==1.3.3
idna==3.4
loguru==0.6.0
multidict==6.0.4
openai==0.27.2
pydantic==1.10.5
PyYAML==6.0
requests==2.28.2
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
